% Generated by roxygen2 (4.0.2): do not edit by hand
\name{abstract.ggplot}
\alias{abstract.ggplot}
\title{Experimental Abstraction Method for GGPlot}
\usage{
\method{abstract}{ggplot}(x, ...)
}
\description{
Not entirely sure this can ever work well since so much of \code{ggplot} is
done with \code{proto} objects and those do not really use meta data, which makes
\code{alike} rather useless.
}
\keyword{internal}

