% Generated by roxygen2 (4.0.1): do not edit by hand
\name{alike.recursive_index_list}
\alias{alike.recursive_full_index_list}
\alias{alike.recursive_index_list}
\alias{alike.recursive_terminal_index_list}
\title{\code{`alike`} methods for recursive indeces}
\arguments{
\item{obj.reference}{R object to use as reference in comparison}

\item{obj}{object to compare}

\item{strict}{set to TRUE so that only indices that are equal pass; this is
mostly so you can get the custom error reporting when they don't.  It's a
bit of a mis-use of alikeness.}

\item{...}{extra arguments for methods}
}
\value{
TRUE if the objects are alike, a one length character vector describing why they are not otherwise
}
\description{
Recursive index lists are considered \code{`alike`} if
every terminal index in \code{`obj`} appears in \code{`obj.reference`},
where appearance includes appearing as a subset.  For example,
\code{`c(1, 2, 1)`} would be considered as appearing in
\code{`c(1, 2, 1, 3)`}.  Additionally, every terminal index
in \code{`obj.reference`} must be matched by a terminal index
\code{`obj`}.
}
\seealso{
alike index_structure
}

