% Generated by roxygen2 (4.0.1): do not edit by hand
\name{all_equal}
\alias{all_equal}
\title{Compare Two Objects}
\usage{
all_equal(target, current, tolerance = .Machine$double.eps, ...)
}
\arguments{
\item{target}{the reference object}

\item{current}{the object to compare to the reference object}

\item{tolerance}{numeric 1 length the allowable differences between the
squares of the  differences between the two numeric objects (note the
default is equivalent to the .Machine$double.eps ^ 0.5 \code{`all.equal`}
uses)}

\item{...}{arguments to pass on to \code{`\link{identical}`}}
}
\value{
logical 1 length
}
\description{
Similar to \code{`\link{all.equal}`}, except that:
\itemize{
  \item It is not an S3 generic, all comparisons are handled within this
    function
  \item numeric objects are compared and if no element is more than
    \code{`tolerance`} apart the objects are considered equal
  \item other objects are compared with \code{`\link{identical}`}
  \item returns \code{FALSE} on failure, not a string explaining the error
  \item does not check attributes if both \code{`target`} and \code{`current`}
    are numeric; implicitly does so if they are not as it uses
    \code{`identical`} in that case.
}
The main purpose of this function is to allow reasonable comparisons of
numerics with a little less overhead than \code{`\link{all.equal}`}.
}

