% Generated by roxygen2 (4.0.2): do not edit by hand
\name{attr_compare}
\alias{attr_compare}
\alias{class_compare,}
\alias{dim_compare}
\alias{dimname_compare,}
\alias{name_compare,}
\title{Compare Attributes}
\usage{
attr_compare(target, current, attr.mode = 0L)
}
\arguments{
\item{int_mode}{}
}
\description{
R interface for an internal C functions used by \code{`alike`}.  Provided
primarily for unit testing purposes
}
\keyword{internal}

