% Generated by roxygen2 (4.0.1): do not edit by hand
\name{index_structure_c}
\alias{index_structure_c}
\title{A C version of Index Structure}
\usage{
index_structure_c(x, term_nodes_only)
}
\arguments{
\item{x}{a list}

\item{terminal.nodes.only}{a logical vector}
}
\description{
A C version of Index Structure
}

