% Generated by roxygen2 (4.0.1): do not edit by hand
\name{index_structure_common}
\alias{index_structure_common}
\title{Helper function for index_structure}
\usage{
index_structure_common(x, terminal.nodes.only = FALSE, recurse.include,
  recurse.exclude = function(y) FALSE, ...)
}
\arguments{
\item{recurse.include}{1 parameter function that determines whether an object is a recursive
element and should be recursed into (e.g., use \code{`\link{is.list}`}) to recurse into
list like objects.}

\item{x}{the recursive object to generate indeces for}

\item{terminal.nodes.only}{set to FALSE if you want all intermediate recursive
objects in addition to the "leaves" of the recursive structure (see examples).
Be careful with this setting if you are dealing with a recursive structure
containing large objects as this could lead to the large object being copied
many, many times over.}

\item{recurse.exclude}{function that will evaluate whether an element of \code{`x`}
should be recursed into or not, should accept only one argument}

\item{...}{included for compatibility with future methods}
}
\description{
Only intended for use by other functions in package.
}
\details{
Currently used for \code{`index_structure.list`} and
\code{`index_structure.call`}
}
\keyword{internal}

