% Generated by roxygen2 (4.0.1): do not edit by hand
\name{is.recursive_index}
\alias{is.recursive_full_index_list}
\alias{is.recursive_index}
\alias{is.recursive_index_list}
\alias{is.recursive_terminal_index_list}
\title{Test Whether Objects Are Recursive Index Objects}
\usage{
is.recursive_index(x)
}
\arguments{
\item{x}{object to test}
}
\value{
logical 1 length TRUE if object is what is expected, FALSE otherwise
}
\description{
Test Whether Objects Are Recursive Index Objects
}
\seealso{
\code{`\link{index_structure}`}
}

