% Generated by roxygen2 (4.0.2): do not edit by hand
\docType{data}
\name{MachDblEpsSqrt}
\alias{MachDblEpsSqrt}
\title{Pre-calculated Precision Level}
\format{\preformatted{ num 1.49e-08
}}
\usage{
MachDblEpsSqrt
}
\description{
Used to limit overhead of calls the require use of \code{`.Machine$double.eps ^ 0.5`}
}
\keyword{datasets}

