% Generated by roxygen2 (4.0.1): do not edit by hand
\name{terminal_indices}
\alias{terminal_indices}
\title{Remove Non-Terminal Indeces From Index List}
\usage{
terminal_indices(index.list)
}
\arguments{
\item{index.list}{list of recursive index vectors}
}
\value{
recursive_terminal_index_list
}
\description{
Any recursive indices that are fully contained in another index in the index
list (starting with the first element of this other index) are by definition
non-terminal, and are removed.  For example, in:
\preformatted{
list(
  c(1, 2, 3),
  c(1, 2, 3, 4),
  c(2, 1, 2, 3, 4)
)
}
the first recursive index is fully contained in the second, so it is not a
terminal index.  However, the second index is terminal because the \bold{starting}
sequence of the third index doesn't fully contain it.
}
\seealso{
\code{`\link{index_structure}`}
}

