% Generated by roxygen2 (4.0.2): do not edit by hand
\name{type_alike}
\alias{.type_alike}
\alias{type_alike}
\title{Compare Types of Objects}
\usage{
type_alike(target, current, mode = 0L, tolerance = MachDblEpsSqrt)
}
\arguments{
\item{target}{the object to test type alikeness against}

\item{current}{the object to test the type alikeness of}

\item{mode}{integer(1L) in 0:2, see details}

\item{tolerance}{see \code{`tolerance`} parameter for \code{`\link{type_of}`}}
}
\description{
By default, checks \code{`\link{type_of}`} objects and two objects are
considered \code{`type_alike`} if they have the same type.  There is special
handling for integers and reals, whereby if \code{`current`} integer or
integer-like (e.g. 1.0) it will match real or integer \code{`target`}
values.
}
\details{
Specific behavior can be tuned with the \code{`mode`} parameter the values
of which range from \code{`0L`} to \code{`2L`}, with a lower value
corresponding to more relaxed comparison level.

\itemize{
  \item 0: integer like reals (e.g. \code{`1.0`}) can match against integer
    templates, and integers always match real templatesf (default)
  \item 1: integers always match against numeric templates, but not vice
    versa, and integer-like reals are treated only as reals
  \item 2: integers only match integer templates, and numerics only match
    numeric templates
}
}
\seealso{
type_of, alike
}

