% Generated by roxygen2 (4.0.2): do not edit by hand
\name{type_of}
\alias{.type_of}
\alias{type_of}
\title{Similar to \code{`\link{typeof}`}, but Treats Numerics Differently}
\usage{
type_of(object, tolerance = MachDblEpsSqrt)
}
\arguments{
\item{object}{the object to check the type of}

\item{tolerance}{see details}
}
\value{
character(1L) the type of the object
}
\description{
Numerics that look like integers at tolerance \code{`tolerance`} are reported
as integers.  Otherwise the same as \code{`typeof`}.
}
\details{
\code{`tolerance`} controls what is considered "integer-like". "integer-likeness"
is roughly defined as occurring when \code{`all.equal(as.integer(x), x) == TRUE`}.
The \code{`tolerance`} value corresponds to the value of the \code{`tolerance`}
argument used by \code{`all.equal.numeric`}.  Note though this is only an
approximate comparison as \code{`type_of`} does not use \code{`all.equal`}.
The default tolerance value is equal to \code{`.Machine$double.eps ^ .5`},
though note that this value is pre-computed when the package is loaded and stored in
\code{`alike:::MachDblEpsSqrt`} in order to minimize function overhead.

\code{`.typeof`} is a slightly faster version that does not allow you to
modify the \code{`tolerance`} parameter.
}
\examples{
type_of(1.0001)                     # numeric
type_of(1 + 1e-20)                  # integer
type_of(1)                          # integer
type_of(1, tolerance=1L)            # numeric
type_of(data.frame(a=1:3))          # list
}

