% Generated by roxygen2 (4.0.2): do not edit by hand
\name{attr_compare}
\alias{attr_compare}
\alias{class_compare}
\alias{dep_alike}
\alias{dim_compare}
\alias{dimname_compare}
\alias{env_track}
\alias{fun_alike}
\alias{lang_alike}
\alias{match_call_alike}
\alias{name_compare}
\alias{ts_compare}
\title{Internal Funs}
\usage{
attr_compare(target, current, attr.mode = 0L)
}
\arguments{
\item{int_mode}{}
}
\description{
R interface for an internal C functions used by \code{alike}.  Provided
primarily for unit testing purposes
}
\keyword{internal}

