% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/misc.R
\docType{data}
\name{MachDblEpsSqrt}
\alias{MachDblEpsSqrt}
\title{Pre-calculated Precision Level}
\format{\preformatted{ num 1.49e-08
}}
\usage{
MachDblEpsSqrt
}
\description{
Used to limit overhead of calls the require use of
\code{.Machine$double.eps ^ 0.5}
}
\keyword{datasets}

